+++ CHANGELOG.md
@@ -6,7 +6,9 @@ Versions follow [CalVer](https://calver.org).
 
 ### Added
 
-- Add a ``Mesh`` class to build multidimensional meshes.
+- Add `size` property to `gspits.mesh.Partition` class.
+- Ass `duration` property to `gspits.mesh.TimePartition` class.
+- Add a `Mesh` class to build multidimensional meshes.
 - Add `sphinx-book-theme` theme.
 - Install `myst-parser` package so we can write documentation using markdown.